Ensure matplot window is closed at program exit and stabilize
Previously, the matplot window wasn't closed down correctly at
program exit. Resizing the window to max is not working on mac
osx. Here both issues are fixed and we wait for the user after
showing the window. Moreover, the EOF exception is avoid by
using raw_input in stead of input.